end-to-end-applications/python/chat-bot/requirements.txt
restate_sdk==0.1.1
hypercorn
requests
slack_sdk